Speech_To_Text (Offline Using Openai's Whisper)/From an audio file/requirement.txt
openai-whisper==20240930
ffmpeg
